combined backend with henry and fixed routes
