+++ back-end/package.json
@@ -5,7 +5,7 @@
   "main": "app.jsx",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "nodemon index.js"
+    "start": "nodemon app.jsx"
   },
   "keywords": [],
   "author": "",
